+++ package.js
@@ -32,18 +32,13 @@ function configure( api ){
     _use( 'check' );
     _use( 'ecmascript' );
     _use( 'less@4.0.0', 'client' );
+    _use( 'pwix:date@1.0.0-rc' );
     _use( 'pwix:ui-bootstrap5@2.0.0' );
     _use( 'pwix:ui-fontawesome6@1.0.0' );
     _use( 'pwix:ui-utils@1.2.0' );
     _use( 'random' );
     _use( 'reactive-var' );
     _use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
-
-    _use( 'pwix:bootbox@1.5.0' );
-    _use( 'pwix:i18n@1.5.0' );
-    _use( 'pwix:modal@2.0.0' );
-    _use( 'pwix:tabbed@1.0.0' );
-    _use( 'tracker' );
     api.addFiles( 'src/client/components/DateInput/DateInput.js', 'client' );
 }